Refactor per gemini review.
