+++ Sources/MutationRef.swift
@@ -20,26 +20,8 @@ struct MutationRequest<Variable: OperationVariable>: OperationRequest {
   private(set) var variables: Variable?
 
   // Computed requestId
-  lazy var requestId: String = {
-    var keyIdData = Data()
-    if let nameData = operationName.data(using: .utf8) {
-      keyIdData.append(nameData)
-    }
-
-    if let variables {
-      let encoder = JSONEncoder()
-      encoder.outputFormatting = .sortedKeys
-      do {
-        let jsonData = try encoder.encode(variables)
-        keyIdData.append(jsonData)
-      } catch {
-        DataConnectLogger
-          .warning("Error encoding variables to compute request identifier: \(error)")
-      }
-    }
-
-    return keyIdData.sha256String
-  }()
+  lazy var requestId: String = OperationUtils
+    .computeRequestId(operationName: operationName, variables: variables)
 
   init(operationName: String, variables: Variable? = nil) {
     self.operationName = operationName
