+++ Sources/Queries/QueryRequest.swift
@@ -22,26 +22,8 @@ struct QueryRequest<Variable: OperationVariable>: OperationRequest, Hashable, Eq
   private(set) var variables: Variable?
 
   // Computed requestId
-  lazy var requestId: String = {
-    var keyIdData = Data()
-    if let nameData = operationName.data(using: .utf8) {
-      keyIdData.append(nameData)
-    }
-
-    if let variables {
-      let encoder = JSONEncoder()
-      encoder.outputFormatting = .sortedKeys
-      do {
-        let jsonData = try encoder.encode(variables)
-        keyIdData.append(jsonData)
-      } catch {
-        DataConnectLogger
-          .warning("Error encoding variables to compute request identifier: \(error)")
-      }
-    }
-
-    return keyIdData.sha256String
-  }()
+  lazy var requestId: String = OperationUtils
+    .computeRequestId(operationName: operationName, variables: variables)
 
   init(operationName: String, variables: Variable? = nil) {
     self.operationName = operationName
